Move findTranslations logic here instad of in portal
