+++ .storybook/decorators/i18nDecorator.js
@@ -2,13 +2,13 @@ import React from "react"
 import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { I18nextProvider, initReactI18next } from 'react-i18next';
-import { translations }  from './../../lib/i18n'
+import { resources }  from './../../lib/i18n'
 
 i18n
   .use(LanguageDetector)
   .use(initReactI18next)
   .init({
-      resources: translations,
+      resources,
       lng: 'sv',
       fallbackLng: 'sv',
       whitelist: ['en', 'sv'],
