+++ lib/i18n/index.js
@@ -1,6 +1,12 @@
 import i18next from 'i18next';
 import addTranslation from './addTranslation';
-import resources from './translationResources';
+import findTranslations from './findTranslations';
+
+/**
+ * Find and build all our translations
+ */
+const context = require.context('./../../src/', true, /\.translation.js/);
+const resources = findTranslations(context, 'asurgentui');
 
 /**
  * Run after i18n has been initated
@@ -15,5 +21,6 @@ export {
   i18next,
   resources,
   addTranslation,
+  findTranslations,
   addComponentTranslations,
 };
